backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-cloud-storage
google-cloud-documentai
google-cloud-firestore
google-cloud-aiplatform
langchain
langgraph
langchain-google-vertexai
